// src/HOPPER.Frontend/src/app/api/model/modUploadResultDto.ts
/**
 * HOPPER.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ModDto } from './modDto';
import { FailedUploadDto } from './failedUploadDto';


export interface ModUploadResultDto { 
    uploaded: Array<ModDto>;
    failed: Array<FailedUploadDto>;
}